Add in code documentation for how mutations work
I just got started on learning about mutations and came across this sample repository. It would help a lot if the code is documented for newcomers to understand what is going on here.

Also it'd be pretty helpful to have some sample queries to play around with in Graphiql. So I came up with a sample query and also added in some comments to what is going on at each section of the code. 

I'm not very good with this yet, but this was how I understood the code. Feel free to let me know if any of it is wrong, I'm still learning :)